+++ services/web/mgmt/2020-12-01/web/enums.go
@@ -1211,6 +1211,23 @@ func PossibleResolveStatusValues() []ResolveStatus {
 	return []ResolveStatus{AccessToKeyVaultDenied, FetchTimedOut, Initialized, InvalidSyntax, MSINotEnabled, OtherReasons, Resolved, SecretNotFound, SecretVersionNotFound, UnauthorizedClient, VaultNotFound}
 }
 
+// ResourceNotRenewableReason enumerates the values for resource not renewable reason.
+type ResourceNotRenewableReason string
+
+const (
+	// ExpirationNotInRenewalTimeRange ...
+	ExpirationNotInRenewalTimeRange ResourceNotRenewableReason = "ExpirationNotInRenewalTimeRange"
+	// RegistrationStatusNotSupportedForRenewal ...
+	RegistrationStatusNotSupportedForRenewal ResourceNotRenewableReason = "RegistrationStatusNotSupportedForRenewal"
+	// SubscriptionNotActive ...
+	SubscriptionNotActive ResourceNotRenewableReason = "SubscriptionNotActive"
+)
+
+// PossibleResourceNotRenewableReasonValues returns an array of possible values for the ResourceNotRenewableReason const type.
+func PossibleResourceNotRenewableReasonValues() []ResourceNotRenewableReason {
+	return []ResourceNotRenewableReason{ExpirationNotInRenewalTimeRange, RegistrationStatusNotSupportedForRenewal, SubscriptionNotActive}
+}
+
 // ResourceScopeType enumerates the values for resource scope type.
 type ResourceScopeType string